equiv: tier-9 generic helper model, driven by the UAPI prototypes
The helper tail had the same shape as tier 8's kfunc tail — ~106 sites
over ~30 helpers with no dominant member — so it gets the same treatment:
one generic model instead of thirty bespoke ones. The prototypes come
from the kernel's OWN uapi header (bpfhelpers.py parses the doc comments
above the FN() list, 184 signatures), so what the model believes a helper
reads is the kernel's declaration, not a hand-maintained table that can
drift.

Each argument enters the trace event the way the kernel reads it:
scalars at their declared width, pointers into observable regions by
identity (their contents are already compared in their own right, and the
prototype's pointee is often not the layout the program sees — the
context argument is declared `struct xdp_buff *` where the program holds
an `xdp_md`), private buffers by their bytes. Extents come from the
prototype too: struct pointees sized against the same kernel BTF used for
CO-RE, and `void *dst, u32 size` pairs sized from the paired argument,
which the header states positionally and by name. A `void *` with no
length partner, and a symbolic length, still bail — the extent is
precisely the thing not worth guessing. 6 helper bail-sites remain, from
106.

Three model refinements fell out of triage, each now a hermetic test:

- A private buffer's ADDRESS is not an observable. The two objects place
  their locals at different frame offsets, and even in different FRAMES —
  clang inlined clobber_regs_stack where rustc left it a bpf2bpf callee,
  making the same local `stack:T` in one object and `stack:T:f1` in the
  other. Identity is now the aliasing structure — is this argument the
  same buffer as an earlier one — and nothing else.

- Bytes the program never stored to are reported as zero rather than as
  residue. An output buffer holds nothing else at call time, and each
  object would otherwise read its own; it also makes an uninitialized C
  local agree with the `[0u8; N]` its translation declares. Cost: a byte
  stored as zero reads the same as one never stored.

- Writtenness is RECORDED at store time, not inferred by comparing
  against the entry value. Residue is symbolic, so `cur == init` is
  satisfiable for a byte that was written, and a solver picking that
  model on one side only invented a divergence between two buffers
  holding identical data. That first attempt looked right and was wrong;
  the sweep found it.

550 objects: 1274 EQUIV / 0 INEQUIV / 17 WAIVED, 356 fully proved (from
322). Five real translation bugs, all QEMU-gated:

- test_xdp_meta kept bpf_xdp_get_buff_len()'s u64 where C narrows to
  `__u32 len` before the length guard.
- test_tc_neigh narrowed skb->protocol to u16 where C switches on the
  full __u32.
- ima narrowed a `bool` ctx argument to its low byte where C's bool
  conversion tests all 64 bits, and used Rust `bool` for four globals
  clang compiles three different ways in this one file (`!= 0`, `!= 1`,
  and `(x & 1) != 0`) — the sharpest bool-global case after
  test_sockmap_listen.
- tailcall_bpf2bpf_hierarchy3 was missing the clobber_regs_stack() call
  outright.
- syscall named its function-local statics in SCREAMING_CASE, so they
  never paired with C's by name.

Three waived with diagnosis: bpf_iter_task_btf and rcu_read_lock both
take the C source's own fallback path (no rustc equivalent for
__builtin_btf_type_id; the other path is verifier-rejected before it can
run), and syscall's remainder is the right-sized prefix structs its
header documents against C's full 168-byte `union bpf_attr` statics.

+++ progs/test_tc_neigh.rs
@@ -144,11 +144,13 @@ fn is_remote_ep_v6(skb: *const __sk_buff, addr: &[u8; 16]) -> bool {
 
 #[inline(always)]
 fn tc_redir(skb: *const __sk_buff, v4_addr: u32, v6_addr: &[u8; 16], ifindex: u32) -> i32 {
-    let protocol = vload!((*skb).protocol) as u16;
+    // C switches on the full `__u32 skb->protocol`; narrowing to u16 here
+    // would match on the low half alone.
+    let protocol = vload!((*skb).protocol);
 
-    let redirect = if protocol == htons(ETH_P_IP) {
+    let redirect = if protocol == htons(ETH_P_IP) as u32 {
         is_remote_ep_v4(skb, htonl(v4_addr))
-    } else if protocol == htons(ETH_P_IPV6) {
+    } else if protocol == htons(ETH_P_IPV6) as u32 {
         is_remote_ep_v6(skb, v6_addr)
     } else {
         false
